refactor(script): split monster buildins into buildin_mob.c
Issue #61 sixth step: move monster/guardian/mob-skill helpers into
buildin_mob.c; replace getmapmoblist str_buf access with get_str().
